Add filter to blocked cards

+++ CW/CW/Services/PatternService.cs
@@ -30,7 +30,7 @@ namespace CW.Services
             }
         }
 
-        public async Task<string> CreatePattern(string pattern_name, string from, string to, int amount)
+        public async Task<string> CreatePattern(string pattern_name, string from, string to, double amount)
         {
             string json = await _client.create_pattern(App.GetUser().login, pattern_name, from, to, amount);
             try
